requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile
#
--index-url https://pypi.tuna.tsinghua.edu.cn/simple
--trusted-host https

addict==2.4.0
    # via
    #   mmcv
    #   mmcv-full
alembic==1.4.3
    # via -r requirements.in
amqp==5.1.1
    # via kombu
astor==0.8.1
    # via paddlepaddle
async-timeout==4.0.2
    # via redis
billiard==3.6.4.0
    # via celery
celery==5.2.7
    # via -r requirements.in
certifi==2023.5.7
    # via requests
charset-normalizer==2.0.12
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   flask
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
csaps==1.1.0
    # via -r requirements.in
cycler==0.11.0
    # via matplotlib
decorator==5.1.1
    # via paddlepaddle
dependency-injector==4.39.1
    # via -r requirements.in
deprecated==1.2.13
    # via
    #   limits
    #   redis
et-xmlfile==1.1.0
    # via openpyxl
flask==1.1.2
    # via
    #   -r requirements.in
    #   flask-limiter
flask-limiter==1.5
    # via -r requirements.in
fonttools==4.39.4
    # via matplotlib
greenlet==2.0.2
    # via sqlalchemy
idna==3.3
    # via
    #   -r requirements.in
    #   requests
imageio==2.15.0
    # via
    #   -r requirements.in
    #   imgaug
    #   scikit-image
imgaug==0.4.0
    # via -r requirements.in
importlib-metadata==6.6.0
    # via numba
importlib-resources==5.12.0
    # via limits
itsdangerous==2.1.2
    # via flask
jinja2==2.11.2
    # via
    #   -r requirements.in
    #   flask
joblib==1.2.0
    # via -r requirements.in
kiwisolver==1.4.4
    # via matplotlib
kombu==5.2.4
    # via
    #   -r requirements.in
    #   celery
limits==3.4.0
    # via flask-limiter
llvmlite==0.39.1
    # via numba
mako==1.2.4
    # via alembic
markupsafe==2.1.2
    # via
    #   jinja2
    #   mako
matplotlib==3.5.1
    # via
    #   -r requirements.in
    #   imgaug
    #   mmdet
    #   pycocotools
    #   seaborn
mmcv==1.4.0
    # via -r requirements.in
mmcv-full==1.4.0
    # via -r requirements.in
mmdet==2.25.2
    # via -r requirements.in
mpi4py==3.0.3
    # via -r requirements.in
networkx==3.1
    # via scikit-image
numba==0.56.2
    # via -r requirements.in
numpy==1.23.3
    # via
    #   -r requirements.in
    #   csaps
    #   imageio
    #   imgaug
    #   matplotlib
    #   mmcv
    #   mmcv-full
    #   mmdet
    #   numba
    #   opencv-python
    #   opt-einsum
    #   paddlepaddle
    #   pandas
    #   pycocotools
    #   pywavelets
    #   scikit-image
    #   scipy
    #   seaborn
    #   tifffile
opencv-python==4.5.5.62
    # via
    #   -r requirements.in
    #   imgaug
    #   mmcv
    #   mmcv-full
openpyxl==3.0.9
    # via -r requirements.in
opt-einsum==3.3.0
    # via paddlepaddle
packaging==21.3
    # via
    #   -r requirements.in
    #   limits
    #   matplotlib
    #   mmcv
    #   mmcv-full
    #   redis
    #   scikit-image
paddle-bfloat==0.1.7
    # via paddlepaddle
paddlepaddle==2.4.0rc0
    # via -r requirements.in
pandas==1.4.1
    # via
    #   -r requirements.in
    #   seaborn
pillow==9.0.1
    # via
    #   -r requirements.in
    #   imageio
    #   imgaug
    #   matplotlib
    #   mmcv
    #   mmcv-full
    #   paddlepaddle
    #   scikit-image
prompt-toolkit==3.0.38
    # via click-repl
protobuf==3.20.0
    # via paddlepaddle
pyclipper==1.3.0.post2
    # via -r requirements.in
pycocotools==2.0.6
    # via mmdet
pydantic==1.10.7
    # via -r requirements.in
pyjwt==1.6.4
    # via -r requirements.in
pymysql==0.10.1
    # via -r requirements.in
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via
    #   alembic
    #   matplotlib
    #   pandas
python-editor==1.0.4
    # via alembic
pytz==2023.3
    # via
    #   celery
    #   pandas
pywavelets==1.4.1
    # via scikit-image
pyyaml==6.0
    # via
    #   mmcv
    #   mmcv-full
pyzbar==0.1.9
    # via -r requirements.in
redis==4.3.4
    # via -r requirements.in
requests==2.27.1
    # via
    #   -r requirements.in
    #   paddlepaddle
scikit-image==0.19.1
    # via
    #   -r requirements.in
    #   imgaug
scipy==1.8.0
    # via
    #   -r requirements.in
    #   csaps
    #   imgaug
    #   scikit-image
seaborn==0.12.0
    # via -r requirements.in
shapely==1.8.0
    # via
    #   -r requirements.in
    #   imgaug
six==1.16.0
    # via
    #   -r requirements.in
    #   click-repl
    #   dependency-injector
    #   flask-limiter
    #   imgaug
    #   mmdet
    #   paddlepaddle
    #   python-dateutil
sqlalchemy==1.4.22
    # via
    #   -r requirements.in
    #   alembic
terminaltables==3.1.10
    # via mmdet
tifffile==2022.2.2
    # via
    #   -r requirements.in
    #   scikit-image
tomli==2.0.1
    # via yapf
tqdm==4.64.1
    # via -r requirements.in
typing-extensions==4.5.0
    # via
    #   limits
    #   pydantic
urllib3==1.26.15
    # via requests
uwsgi==2.0.20
    # via -r requirements.in
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via prompt-toolkit
werkzeug==1.0.1
    # via
    #   -r requirements.in
    #   flask
wrapt==1.15.0
    # via deprecated
xlsxwriter==3.0.2
    # via -r requirements.in
xmltodict==0.13.0
    # via -r requirements.in
yapf==0.33.0
    # via
    #   mmcv
    #   mmcv-full
zipp==3.15.0
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
